services/auth-service/requirements.txt
flask==2.3.0
pyjwt==2.8.0
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0